% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blsdata.R
\docType{data}
\name{blsdata}
\alias{blsdata}
\title{Dataset from the Box Lunch Study, taken from the "visTree" package.}
\format{
An object of class \code{data.frame} with 226 rows and 26 columns.
}
\source{
\href{https://github.com/AshwiniKV/visTree/tree/master/data}{https://github.com/AshwiniKV/visTree/tree/master/data}
}
\usage{
data(blsdata)
}
\description{
Information
}
\examples{
data(blsdata)
}
\references{
(\href{https://github.com/AshwiniKV/visTree}{https://github.com/AshwiniKV/visTree})
}
\keyword{datasets}
