% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_funs.R
\name{correctPVal}
\alias{correctPVal}
\title{Get a valid pvalue for a test}
\usage{
correctPVal(phiInterval, nu, y, sigma)
}
\arguments{
\item{phiInterval}{the selective inference interval}

\item{nu}{the contrast vector for the hypothesis test}

\item{y}{the response data y}

\item{sigma}{the (assumed known) noise standard deviation}
}
\value{
a p-value.
}
\description{
Based on the similar function from the Outference package,
which accomplishes a similar task. Code modified from code found at
https://github.com/shuxiaoc/outference
}
