% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_tree_inference_mat.R
\name{fullTreeInference}
\alias{fullTreeInference}
\title{Constructs a large matrix storing a pvalue for every split and a CI for every node.
Used as a precursor to plotting.}
\usage{
fullTreeInference(tree, sigma_y = sd(tree$model[, 1]))
}
\arguments{
\item{tree}{An rpart object. Must have been built with rpart arguement "model=TRUE".}

\item{sigma_y}{The known error variance. If not provided, is estimated with a conservative guess.}
}
\value{
A large matrix storing lots of pvalues and confidence intervals.
}
\description{
Constructs a large matrix storing a pvalue for every split and a CI for every node.
Used as a precursor to plotting.
}
