% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_helpers_rpart_plot.R
\name{get.default.extra}
\alias{get.default.extra}
\title{This is the type of thing I think I can probably delete.}
\usage{
get.default.extra(obj, class.stats)
}
\description{
This is the type of thing I think I can probably delete.
}
