% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_funs.R
\name{getAllBranches}
\alias{getAllBranches}
\title{Get a list of branches in tree to easily find out what you want to do inferenceon}
\usage{
getAllBranches(tree)
}
\arguments{
\item{tree}{An rpart object. Must have been built with model=TRUE}
}
\description{
Get a list of branches in tree to easily find out what you want to do inferenceon
}
