% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_KL_helpers.R
\name{getDesc}
\alias{getDesc}
\title{Returns vector of descendants in ``tree`` for node ``nn``.}
\usage{
getDesc(tree, nn)
}
\arguments{
\item{tree}{an rpart object.}

\item{nn}{the number of a node in tree. Must correspond to a valid node number (found in row names of tree$frame).}
}
\value{
a vector of node numbers corresponding to the descendants of ``nn`` in ``tree``.
}
\description{
Returns vector of descendants in ``tree`` for node ``nn``.
}
