% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi_interval_funs.R
\name{getInterval_full}
\alias{getInterval_full}
\title{Workhorse function for computing conditioning sets.}
\usage{
getInterval_full(tree, nu, branch, sib = FALSE, grow = FALSE, prune = FALSE)
}
\arguments{
\item{tree}{An rpart object. Must have been built with model=TRUE arguement unless X and y are provided separately}

\item{nu}{The contrast vector that defines the parameter.}

\item{branch}{A vector of strings specifying the branch.}

\item{sib}{If you are doing inference and nu=nu_sib, this takes advantage of some computational speedups!}

\item{grow}{Set this to true if you only want to compute S_grow}

\item{prune}{Set this to true if you only want to see what gets REMOVED from S_grow during pruning}
}
\value{
an object of class Interval that defines the set S.
}
\description{
Returns the interval for phi such that Tree(y'(phi)) (where the definition of y'(phi) depends on nu)
contains the series of branch "branch". An rpart object build with model=TRUE must be provided.
}
