% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi_interval_CP.R
\name{getInterval_prune}
\alias{getInterval_prune}
\title{Workhorse function for computing conditioning sets; prune only!!}
\usage{
getInterval_prune(base_tree, nu, splits)
}
\arguments{
\item{base_tree}{An rpart object. Must have been built with model=TRUE arguement unless X and y are provided separately}

\item{nu}{The contrast vector that defines the parameter.}

\item{splits}{A vector of strings specifying the splits.}
}
\value{
an object of class Interval that defines the set S.
}
\description{
Returns the interval for phi such that Tree(y'(phi)) (where the definition of y'(phi) depends on nu)
contains the series of splits "splits".Either base_tree with model=TRUE or X and y must be provided, not both.
}
