% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_int_funs.R
\name{isSameIntervals}
\alias{isSameIntervals}
\title{Comparison between two intervals}
\usage{
isSameIntervals(int1, int2)
}
\arguments{
\item{int1, int2}{"Intervals" objects.}
}
\value{
This function returns the desired logical result.
}
\description{
This functions returns \code{TRUE} if and only if two intervals are the same.
}
\keyword{internal}
