% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_int_funs.R
\name{magicfun}
\alias{magicfun}
\title{A helper function for approximating normal tail probabilities
This comes directly from the Outference Package}
\usage{
magicfun(z)
}
\arguments{
\item{z, }{the number where the function is evaluated.}
}
\value{
This function returns the value of the function evaluated at \code{z}.
}
\description{
For \eqn{Z ~ N(0, 1)}, we have the approximation
    \eqn{P(Z \ge z) \approx }\code{magicfun(z)*exp(-z^2/2)}.
}
\references{
Bryc, Wlodzimierz. "A uniform approximation to the right normal tail integral."
    Applied mathematics and computation 127.2 (2002): 365-374.
}
