% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_funs.R
\name{nodeInference}
\alias{nodeInference}
\title{Get a confidence interval for the mean response in a terminal node.}
\usage{
nodeInference(tree, node, sigma_y = NULL, alpha = 0.05)
}
\arguments{
\item{tree}{An rpart object. The tree must have been built with arguement model=TRUE,}

\item{node}{identify the node!!. Need to explain this. Must be terminal node.}

\item{sigma_y}{enter the known noise SD (for now)}

\item{alpha}{The significance level. We build a 1-alpha CI.}
}
\value{
a list with 3 things.
}
\description{
Get a confidence interval for the mean response in a terminal node.
}
