% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_funs.R
\name{splitInference}
\alias{splitInference}
\title{Get inference for a difference in means between two terminal nodes}
\usage{
splitInference(tree, locTest, sigma_y = NULL, alpha = 0.05)
}
\arguments{
\item{tree}{an rpart object. Must have been built with model=TRUE.}

\item{locTest}{identify the pair of nodes that you are testing for. this should be a vector of length 2. the numbers should correspond to entries in tree$where}

\item{sigma_y}{enter the known SD of y. If none is entered, the default is to use the total SST of y,
which is conservative.}
}
\description{
Get inference for a difference in means between two terminal nodes
}
