% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_utils.R
\name{summary.branch_inference}
\alias{summary.branch_inference}
\title{Summary of inference for a branch in a tree.}
\usage{
\method{summary}{branch_inference}(x, ...)
}
\arguments{
\item{x}{branch_inference object}
}
\description{
Summary of inference for a branch in a tree.
}
