% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_KL_helpers.R
\name{tree_KL}
\alias{tree_KL}
\title{Returns the tree that remains after all nodes in "tree" BESIDES the ancestors of node 'nn' have been visited
for bottom-up pruning with parameter lambda.}
\usage{
tree_KL(tree, lambda, nn, y = NULL)
}
\arguments{
\item{tree}{An rpart object. If y is not provided, must have been built with model=TRUE}

\item{lambda}{The cost complexity parameter for pruning.}

\item{nn}{the node number of the region whose ancestors you want to leave in the tree.}

\item{y}{the response vector to evaluate the pruning on.}
}
\value{
An rpart object; intermediately pruned.
}
\description{
Returns the tree that remains after all nodes in "tree" BESIDES the ancestors of node 'nn' have been visited
for bottom-up pruning with parameter lambda.
}
