% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{treeval.plot}
\alias{treeval.plot}
\title{This function allows users to plot their rpart object with all splits labeled with p-values and all nodes
labeled with confidence intervals.}
\usage{
treeval.plot(
  tree,
  sigma_y = NULL,
  nn = TRUE,
  printn = 1,
  inferenceType = 2,
  digits = 3,
  alpha = 0.05,
  permute = FALSE,
  ...
)
}
\arguments{
\item{tree}{An rpart tree. Note that the tree must have been built with arguements maxcompete=0 and maxsurrogates=0.}

\item{sigma_y}{Provide the standard deviation of y, if known. If not provided, the sample standard deviation of y will be used
as a conservative estimate.}

\item{nn}{boolean- would you like node numbers to be printed?}

\item{inferenceType}{An integer specifying which pieces of inference information should be added to the plot. The options
currently available are:}

\item{digits}{Integer- how many digits would you like the text in the plot rounded to.}

\item{...}{Additional arguements are passed on to rpart.plot(). Examples include "cex".}
}
\description{
This function allows users to plot their rpart object with all splits labeled with p-values and all nodes
labeled with confidence intervals.
}
