% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blsdata.R, R/data.R
\docType{data}
\name{blsdata}
\alias{blsdata}
\title{Dataset from the Box Lunch Study, taken from the "visTree" package.}
\format{
A data frame with 226 rows and 26 variables

An object of class \code{data.frame} with 226 rows and 26 columns.

\describe{
  \item{trt}{Treatment}
  \item{sex}{Sex}
  \item{bmi0}{BMI}
  \item{skcal}{Snacking kilo calories.}
  \item{srvgfv0}{Serving size of fruits and vegetables.}
  \item{srvgssb}{Serving size of beverages}
  \item{kcal24h0}{Response variable; kcals consumed in 24 hours.}
  \item{disinhibition}{}
  \item{wanting}{}
     \item{hunger}{}
        \item{liking}{}
           \item{resteating}{}
  \item{edeq01}{}
  \item{edeq02}{}
  \item{edeq13}{}
  \item{edeq14}{}
  \item{edeq15}{}
  \item{edeq22}{}
  \item{edeq23}{}
  \item{edeq25}{}
  \item{edeq26}{}
  \item{cdrsbody0}{Body image}
  \item{weighfreq0}{Weighing frequency}
  \item{freqff}{Fast food frequency}
  \item{age}{Age}
  \item{rrvfood}{Relative reinforcement of food}
}

A data frame with 226 rows and 26 variables

An object of class \code{data.frame} with 226 rows and 26 columns.
}
\source{
\href{https://github.com/AshwiniKV/visTree/tree/master/data}{https://github.com/AshwiniKV/visTree/tree/master/data}

\href{https://github.com/AshwiniKV/visTree/tree/master/data}{https://github.com/AshwiniKV/visTree/tree/master/data}
}
\usage{
data(blsdata)

data(blsdata)

data(blsdata)

data(blsdata)
}
\description{
Information

Information
}
\examples{
data(blsdata)

data(blsdata)
data(blsdata)

data(blsdata)
}
\references{
(\href{https://github.com/AshwiniKV/visTree}{https://github.com/AshwiniKV/visTree})

(\href{https://github.com/AshwiniKV/visTree}{https://github.com/AshwiniKV/visTree})
}
\keyword{datasets}
