% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_funs.R
\name{branchInference}
\alias{branchInference}
\title{Pass in a branch!!!}
\usage{
branchInference(
  tree,
  branch,
  type = "reg",
  alpha = 0.05,
  sigma_y = NULL,
  c = 0,
  computeCI = TRUE
)
}
\arguments{
\item{tree}{An rpart object. Must have been built with model=TRUE}

\item{branch}{A vector of splits.}

\item{type}{Pass in either "reg" or "sib"}

\item{alpha}{The significance level forthe confidence interval}

\item{sigma_y}{the error variance.Assumed known but if not we'll use the sample variance.}
}
\value{
A fitted object!! With summary info
}
\description{
Pass in a branch!!!
}
