% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_int_funs.R
\name{computeCI}
\alias{computeCI}
\title{Compute selective confidence interval for parameter v^T mu based on a truncated normal distribution. A slight modification of code found in the
Outference package, available at https://github.com/shuxiaoc/outference.}
\usage{
computeCI(v, y, sigma = NULL, truncation, alpha = 0.05)
}
\arguments{
\item{v}{the vector that defines the parameter of interest; v^T mu}

\item{y}{the observed response vector}

\item{sigma}{The known noise standard deviation. If unknown, we recommend a conservative estimate. If it
is left blank, we automatically use a conservative estimate.}

\item{truncation, }{the truncation set for the statistic v'y.
Computes a confidence interval for the mean of a truncated normal distribution.}

\item{alpha, }{the significance level.}
}
\value{
This function returns a vector of lower and upper confidence limits.
}
\description{
Compute selective confidence interval for parameter v^T mu based on a truncated normal distribution. A slight modification of code found in the
Outference package, available at https://github.com/shuxiaoc/outference.
}
