% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_funs.R
\name{correctPVal}
\alias{correctPVal}
\title{Get a pvalue for a selective Z-test using a truncated normal distribution.}
\usage{
correctPVal(phiInterval, nu, y, sigma)
}
\arguments{
\item{phiInterval}{the conditioning set (truncation interval). An object of class "Interval", where the rows represent the union of
disjoint intervals on the real line.}

\item{nu}{the vector that defines the parameter of interest. We are testing the hypothesis that nu^T mu = 0.}

\item{y}{the response data y.}

\item{sigma}{the (assumed known) noise standard deviation. The assumption is that y_i ~ N(mu_i, sigma^2).}
}
\value{
a p-value.
}
\description{
Based on the similar function from the Outference package,
which accomplishes a similar task. Code modified from code found at
https://github.com/shuxiaoc/outference.
}
