% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_funs.R
\name{getAllBranches}
\alias{getAllBranches}
\title{This function takes an rpart tree and}
\usage{
getAllBranches(tree)
}
\arguments{
\item{tree}{An rpart object. Must have been built with model=TRUE}
}
\description{
This function takes an rpart tree and
}
