% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_KL_helpers.R
\name{getAnc}
\alias{getAnc}
\title{Returns vector of ancestors in ``tree`` for node ``nn``.}
\usage{
getAnc(tree, nn)
}
\arguments{
\item{tree}{an rpart object.}

\item{nn}{the number of a node in tree. Note that node numbers correspond to the row names of tree$frame.}
}
\value{
a vector of ancestors.
}
\description{
Returns vector of ancestors in ``tree`` for node ``nn``.
}
