% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi_interval_funs.R
\name{getBranch}
\alias{getBranch}
\title{Obtain the list of splits that refine a certain region.
Necessary precursor to calling "getInterval"}
\usage{
getBranch(tree, node)
}
\arguments{
\item{tree}{An rpart object. Must have been built with model=TRUE}

\item{node}{A number identifying the region that you want the ancestors of. Numbers correspond to elements in tree$where.}
}
\value{
A vector of strings describing the splits that define the node in the tree.
}
\description{
Obtain the list of splits that refine a certain region.
Necessary precursor to calling "getInterval"
}
