% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_funs.R
\name{getRegion}
\alias{getRegion}
\title{Pass in a tree and a node number (node number like the type that comes up in rpart.plot!!!)}
\usage{
getRegion(tree, nn)
}
\arguments{
\item{tree}{An rpart object. Must have been built with model=TRUE}

\item{nn}{A node number!! As a string or as an integer is fine}
}
\description{
Pass in a tree and a node number (node number like the type that comes up in rpart.plot!!!)
}
