% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_funs.R
\name{treeval.plot}
\alias{treeval.plot}
\title{This function allows users to plot their rpart object with all splits labeled with p-values and all nodes
labeled with confidence intervals.}
\usage{
treeval.plot(tree, inferenceMatrix = NULL, sigma_y = NULL)
}
\arguments{
\item{tree}{An rpart tree.}

\item{inferenceMatrix}{The corrsponding inference matrix. This matrix should have been build with the "getFullTreeInference" function.}

\item{sigma_y}{Need to provide this if you are not providing inference matrix.}
}
\description{
This function allows users to plot their rpart object with all splits labeled with p-values and all nodes
labeled with confidence intervals.
}
