% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_funs.R
\name{treeval.plot}
\alias{treeval.plot}
\title{This function allows users to plot their rpart object with all splits labeled with p-values and all nodes
labeled with confidence intervals.}
\usage{
treeval.plot(
  tree,
  inferenceMatrix = NULL,
  sigma_y = NULL,
  nn = TRUE,
  extra = 1,
  inferenceType = 2,
  digits = 3,
  ...
)
}
\arguments{
\item{tree}{An rpart tree. Note that the tree must have been built with arguements maxcompete=0 and maxsurrogates=0.}

\item{inferenceMatrix}{Optionally provide an inference matrix, built with ``fullTreeInference()``. If not provided, ``fullTreeInference()``
will be called internally (and is relatively slow).}

\item{sigma_y}{Provide the standard deviation of y, if known. If not provided, the sample standard deviation of y will be used
as a conservative estimate.}

\item{digits}{Integer- how many digits would you like the text in the plot rounded to.}

\item{...}{Additional arguements are passed on to rpart.plot(). Examples include "nn", "extra","digits", and "cex".}
}
\description{
This function allows users to plot their rpart object with all splits labeled with p-values and all nodes
labeled with confidence intervals.
}
